EIP1-4307 - S3Client extension functions
